Implements Band registration and getUserById endpoint.

+++ musikmatch-backend/src/main/java/com/tcc/musikmatch/models/Band.java
@@ -6,6 +6,7 @@ import jakarta.persistence.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -32,10 +33,10 @@ public class Band implements Serializable {
     private String bio;
 
     @Column(nullable = false)
-    private double lat;
+    private BigDecimal lat;
 
     @Column(nullable = false)
-    private double lon;
+    private BigDecimal lon;
 
     @Column(nullable = false)
     private String city;
@@ -76,19 +77,19 @@ public class Band implements Serializable {
         this.bio = bio;
     }
 
-    public double getLat() {
+    public BigDecimal getLat() {
         return lat;
     }
 
-    public void setLat(double lat) {
+    public void setLat(BigDecimal lat) {
         this.lat = lat;
     }
 
-    public double getLon() {
+    public BigDecimal getLon() {
         return lon;
     }
 
-    public void setLon(double lon) {
+    public void setLon(BigDecimal lon) {
         this.lon = lon;
     }
